[#17] Bugfixes
Bugfixes to the Drakefile to work with the different security of
Postgres on reporting.test.akvo.org.
Note that the env variable PGPASSWORD needs the password of the DB_USER
to be able to run the psql commands.

Also a small bugfix to the column types for postgres, using bigint
rather than integer to be able to handle larger ints.

+++ csv_import_and_report/csv_to_table_def.py
@@ -45,7 +45,7 @@ def infer_datatype(date_format, data, column):
 
         try:
             _ = int(s)
-            return "integer", 0
+            return "bigint", 0
         except ValueError:
             pass
         try:
